apis/tenancy: split types.go
