+++ config/crds/tenancy.kcp.dev_clusterworkspaces.yaml
@@ -145,7 +145,7 @@ spec:
                 properties:
                   name:
                     description: name is the name of the ClusterWorkspaceType
-                    pattern: ^[a-z0-9]([a-z0-9-]{0,61}[a-z0-9])?
+                    pattern: ^[a-z]([a-z0-9-]{0,61}[a-z0-9])?
                     type: string
                   path:
                     description: path is an absolute reference to the workspace that
