+++ config/root-phase0/apiresourceschema-clusterworkspacetypes.tenancy.kcp.dev.yaml
@@ -2,7 +2,7 @@ apiVersion: apis.kcp.dev/v1alpha1
 kind: APIResourceSchema
 metadata:
   creationTimestamp: null
-  name: v221111-63fc4478.clusterworkspacetypes.tenancy.kcp.dev
+  name: v221128-9e4040046.clusterworkspacetypes.tenancy.kcp.dev
 spec:
   group: tenancy.kcp.dev
   names:
@@ -86,7 +86,7 @@ spec:
               properties:
                 name:
                   description: name is the name of the ClusterWorkspaceType
-                  pattern: ^[a-z0-9]([a-z0-9-]{0,61}[a-z0-9])?
+                  pattern: ^[a-z]([a-z0-9-]{0,61}[a-z0-9])?
                   type: string
                 path:
                   description: path is an absolute reference to the workspace that
@@ -118,7 +118,7 @@ spec:
                     properties:
                       name:
                         description: name is the name of the ClusterWorkspaceType
-                        pattern: ^[a-z0-9]([a-z0-9-]{0,61}[a-z0-9])?
+                        pattern: ^[a-z]([a-z0-9-]{0,61}[a-z0-9])?
                         type: string
                       path:
                         description: path is an absolute reference to the workspace
@@ -165,7 +165,7 @@ spec:
                     properties:
                       name:
                         description: name is the name of the ClusterWorkspaceType
-                        pattern: ^[a-z0-9]([a-z0-9-]{0,61}[a-z0-9])?
+                        pattern: ^[a-z]([a-z0-9-]{0,61}[a-z0-9])?
                         type: string
                       path:
                         description: path is an absolute reference to the workspace
@@ -197,7 +197,7 @@ spec:
                     properties:
                       name:
                         description: name is the name of the ClusterWorkspaceType
-                        pattern: ^[a-z0-9]([a-z0-9-]{0,61}[a-z0-9])?
+                        pattern: ^[a-z]([a-z0-9-]{0,61}[a-z0-9])?
                         type: string
                       path:
                         description: path is an absolute reference to the workspace
