+++ pkg/client/clientset/versioned/cluster/typed/tenancy/v1alpha1/fake/tenancy_client.go
@@ -48,12 +48,12 @@ func (c *TenancyV1alpha1ClusterClient) ClusterWorkspaces() kcptenancyv1alpha1.Cl
 	return &clusterWorkspacesClusterClient{Fake: c.Fake}
 }
 
-func (c *TenancyV1alpha1ClusterClient) ClusterWorkspaceTypes() kcptenancyv1alpha1.ClusterWorkspaceTypeClusterInterface {
-	return &clusterWorkspaceTypesClusterClient{Fake: c.Fake}
-}
-
 func (c *TenancyV1alpha1ClusterClient) ClusterWorkspaceShards() kcptenancyv1alpha1.ClusterWorkspaceShardClusterInterface {
 	return &clusterWorkspaceShardsClusterClient{Fake: c.Fake}
+}
+
+func (c *TenancyV1alpha1ClusterClient) ClusterWorkspaceTypes() kcptenancyv1alpha1.ClusterWorkspaceTypeClusterInterface {
+	return &clusterWorkspaceTypesClusterClient{Fake: c.Fake}
 }
 
 var _ tenancyv1alpha1.TenancyV1alpha1Interface = (*TenancyV1alpha1Client)(nil)
@@ -72,10 +72,10 @@ func (c *TenancyV1alpha1Client) ClusterWorkspaces() tenancyv1alpha1.ClusterWorks
 	return &clusterWorkspacesClient{Fake: c.Fake, Cluster: c.Cluster}
 }
 
-func (c *TenancyV1alpha1Client) ClusterWorkspaceTypes() tenancyv1alpha1.ClusterWorkspaceTypeInterface {
-	return &clusterWorkspaceTypesClient{Fake: c.Fake, Cluster: c.Cluster}
-}
-
 func (c *TenancyV1alpha1Client) ClusterWorkspaceShards() tenancyv1alpha1.ClusterWorkspaceShardInterface {
 	return &clusterWorkspaceShardsClient{Fake: c.Fake, Cluster: c.Cluster}
+}
+
+func (c *TenancyV1alpha1Client) ClusterWorkspaceTypes() tenancyv1alpha1.ClusterWorkspaceTypeInterface {
+	return &clusterWorkspaceTypesClient{Fake: c.Fake, Cluster: c.Cluster}
 }
