+++ pkg/client/clientset/versioned/cluster/typed/tenancy/v1alpha1/tenancy_client.go
@@ -35,8 +35,8 @@ import (
 type TenancyV1alpha1ClusterInterface interface {
 	TenancyV1alpha1ClusterScoper
 	ClusterWorkspacesClusterGetter
-	ClusterWorkspaceTypesClusterGetter
 	ClusterWorkspaceShardsClusterGetter
+	ClusterWorkspaceTypesClusterGetter
 }
 
 type TenancyV1alpha1ClusterScoper interface {
@@ -58,12 +58,12 @@ func (c *TenancyV1alpha1ClusterClient) ClusterWorkspaces() ClusterWorkspaceClust
 	return &clusterWorkspacesClusterInterface{clientCache: c.clientCache}
 }
 
-func (c *TenancyV1alpha1ClusterClient) ClusterWorkspaceTypes() ClusterWorkspaceTypeClusterInterface {
-	return &clusterWorkspaceTypesClusterInterface{clientCache: c.clientCache}
-}
-
 func (c *TenancyV1alpha1ClusterClient) ClusterWorkspaceShards() ClusterWorkspaceShardClusterInterface {
 	return &clusterWorkspaceShardsClusterInterface{clientCache: c.clientCache}
+}
+
+func (c *TenancyV1alpha1ClusterClient) ClusterWorkspaceTypes() ClusterWorkspaceTypeClusterInterface {
+	return &clusterWorkspaceTypesClusterInterface{clientCache: c.clientCache}
 }
 
 // NewForConfig creates a new TenancyV1alpha1ClusterClient for the given config.
