+++ pkg/client/informers/externalversions/generic.go
@@ -111,10 +111,10 @@ func (f *sharedInformerFactory) ForResource(resource schema.GroupVersionResource
 	// Group=tenancy.kcp.dev, Version=V1alpha1
 	case tenancyv1alpha1.SchemeGroupVersion.WithResource("clusterworkspaces"):
 		return &genericClusterInformer{resource: resource.GroupResource(), informer: f.Tenancy().V1alpha1().ClusterWorkspaces().Informer()}, nil
-	case tenancyv1alpha1.SchemeGroupVersion.WithResource("clusterworkspacetypes"):
-		return &genericClusterInformer{resource: resource.GroupResource(), informer: f.Tenancy().V1alpha1().ClusterWorkspaceTypes().Informer()}, nil
 	case tenancyv1alpha1.SchemeGroupVersion.WithResource("clusterworkspaceshards"):
 		return &genericClusterInformer{resource: resource.GroupResource(), informer: f.Tenancy().V1alpha1().ClusterWorkspaceShards().Informer()}, nil
+	case tenancyv1alpha1.SchemeGroupVersion.WithResource("clusterworkspacetypes"):
+		return &genericClusterInformer{resource: resource.GroupResource(), informer: f.Tenancy().V1alpha1().ClusterWorkspaceTypes().Informer()}, nil
 	// Group=tenancy.kcp.dev, Version=V1beta1
 	case tenancyv1beta1.SchemeGroupVersion.WithResource("workspaces"):
 		return &genericClusterInformer{resource: resource.GroupResource(), informer: f.Tenancy().V1beta1().Workspaces().Informer()}, nil
@@ -158,11 +158,11 @@ func (f *sharedScopedInformerFactory) ForResource(resource schema.GroupVersionRe
 	case tenancyv1alpha1.SchemeGroupVersion.WithResource("clusterworkspaces"):
 		informer := f.Tenancy().V1alpha1().ClusterWorkspaces().Informer()
 		return &genericInformer{lister: cache.NewGenericLister(informer.GetIndexer(), resource.GroupResource()), informer: informer}, nil
-	case tenancyv1alpha1.SchemeGroupVersion.WithResource("clusterworkspacetypes"):
-		informer := f.Tenancy().V1alpha1().ClusterWorkspaceTypes().Informer()
-		return &genericInformer{lister: cache.NewGenericLister(informer.GetIndexer(), resource.GroupResource()), informer: informer}, nil
 	case tenancyv1alpha1.SchemeGroupVersion.WithResource("clusterworkspaceshards"):
 		informer := f.Tenancy().V1alpha1().ClusterWorkspaceShards().Informer()
+		return &genericInformer{lister: cache.NewGenericLister(informer.GetIndexer(), resource.GroupResource()), informer: informer}, nil
+	case tenancyv1alpha1.SchemeGroupVersion.WithResource("clusterworkspacetypes"):
+		informer := f.Tenancy().V1alpha1().ClusterWorkspaceTypes().Informer()
 		return &genericInformer{lister: cache.NewGenericLister(informer.GetIndexer(), resource.GroupResource()), informer: informer}, nil
 	// Group=tenancy.kcp.dev, Version=V1beta1
 	case tenancyv1beta1.SchemeGroupVersion.WithResource("workspaces"):
